agents/email/requirements.txt
sendgrid==6.11.0
psycopg2-binary==2.9.9
pydantic==2.5.2
jinja2==3.1.2
python-dateutil==2.8.2
requests==2.31.0
fastapi==0.109.0
uvicorn==0.27.0
